posix-stack: Introduce wrapped_iovec
When sending a collection of temporary_buffers (nowadays -- via span<>
reference) posix stack needs to convert the buffers into an array of
iovecs for vectorized syscalls, sens the iovecs somehow, and then release
the buffers' memory.

Currnetly only TCP data_sink does so, but the UDP socket implementation
will soon need to do the same. To facilitate the buffers management, here's
the wrapper_iovec helper struct that holds vector of iovecs, gets populated
from span of temporary_buffers and maintains the lifetime of data with the
help of deleters chain.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>