net: Introduce datagram_channel::send(span<temporary_buffer>) method
As a replacement for exising send(net::packet) one. The latter will be
deprecated by next patch.

The virtual datagram_channel_impl::send(net::packet) is just patched to
accept span of buffers, the old one is not kept. The _impl class is not
the part of seastar API, presumably, it's not expected to have any other
inheritants from it in 3rd party code.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>