util: Genralize memory_data_sink
The sink that stores temporary buffers put into it into some external
collection is useful on its own. It's used in tests, that store buffers
in stringstream, it's used by memcache, that stores buffers into a
vector of vectors of buffers, and it's used in Scylla, which accumulates
buffers in a vector decorated with some axuliary data.

This patch introduces basic_memory_data_sink template with Container
parameter. The append_buffers(Container&, span<temporary_buffer>) function
is a requirement for Container type.

The new sink is available for API level 9 (and above) only, not to implement
legacy put() methods. It suits the needs of tests, memcached and Scylla.

The vector_data_sink, that forwards temporary_buffers into net::packet is
deprecated in favor of basic_memory_data_sink<vector> for newer API levels.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>