task: Add sched-group disguising/unveiling
Rename task::_sg member into uintptr and keep scheduling group index
in there. Next patch will keep task* pointer unioned with it, so to tell
which value is in there, add disguising of sched group ID by shifting it
left one bit and setting the zeroth one. Also add a static assertion
that when task pointer will be put there, this bit will not be set.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>