sched: Re-implement task queue as singly-linked list
The list (class task_slist) holds pointer to first element, and the pointer
to the "terminating" pointer to enable constant-time push_back().

The element (class task) shares the "next" pointer with sched group index,
thus living in two states -- queued and inactive. When created it gets into
the latter state and carries sched group index on-board. When add_task is
called, the task becomes queued and the sched group index is thus lost. Later,
when the task is picked for execution, the task_queue restores this index
into "current", naturally. Probably the restoration is not needed, but it's
better to be on the safe side.

The task::set_scheduling_group() is no longer callable for queued tasks, but
the method is protected and is only used in two places. First is coroutine
switch-to, which happens on running (i.e. -- not queued) task. Second is the
shared_future tweaking unresolved shared_state (i.e. -- inactive too).

Also, not to implement the begin()/end()/iterators for the task_slist, the
only place that walks the whole queue is re-implemented with the help of
do_for_each(lambda) helper.

Debug-mode task queue shuffling with constant time is no longer possible and
is temporarily removed, next patch will restore it.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>