coroutines: simplify execute_involving_handle_destruction_in_await_suspend()
execute_involving_handle_destruction_in_await_suspend() auto-recognizes
if a task is passed to it, so it can avoid allocating a new task,
but that causes the caller to have to create an artificial operator().

Simplify by having a separate overload for task*.

The artificial operator()() in try_future is removed, and
run_and_dispose() is marked final to reduce call overhead.

Closes #3128