udp: Make datagram implementations return single temporary_buffer<>
The datagram API converts it into a packet, if the get_data() method is
called (will be deprecated soon)

For posix implementation, this is natural transition, as it only has
single-buffer data from recv() syscall.

For native implentation, data linearization is required, but it's
unlikely that a packet arrives fragmented.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>