util: Introduce intrusive split-list
The split list is the collection of objects that's optimized for
scanning from begin to end, with the ability to add entries from two
ends -- front and back, like std::deque does.

Splitting is needed to let CPU pre-fetch next element while scanning
some current one. With classical linked lists this is difficult, because
in order to prefetch the next element CPU needs to read the next pointer
from the current element, which is being read. Thus the forward scan of
a plain list is serielized. With split lists several adjacent elements
can be prefetched in parallel.

To achieve that, the "first" pointer is "sharded" -- there are F first
pointers (as well as F cached last pointers) and the list is populated
in round-robin manner -- first 0th list is appended, then the 1st, then
the 2nd, ... then the Fth and population wraps around and starts over.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>