test/perf: Add seastar::intrusive_split_list<T> scanning test
Use split-list with the factor of 16 as it shows the nicest results on
my CPU (Ryzen Threadripper 2970WX).

This test beats singly linked list greatly, and works ~2x times slower
on 10k elements that the array of pointers does. But unlike array of
pointers, it doesn't require memory allocation on appending operations.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>