core: move set_current_task() out-of-line
This function is currently an inline one in reactor.hh, which forces all
users to include the huge reactor.hh, for this single method. There are
two such users currently:
* src/core/thread.cc
* coroutine/parallel_for_each.hh

The latter is really bad because it is a header, which means the header
bloat is propagated.

Both include reactor.hh just for this one function. Move this function
out-of-line:
* Move declaration to include/seastar/core/internal/current_task.hh
* Move definition to src/reactor.cc

This allows reducing header bloat at the cost of a out-of-line function
call.