reactor: Restore task queue shuffling
Currently shuffling is performed by swapping a newly-activated task with
some random queued one. Previous patch turned task queue into a singly-linked
list, and picking a random task from it in constant time is impossible.

The new shuffling works run-time -- when a queue picks up next front task
to execute, it rolls the dice and optionally re-queued it into the back of
the queue.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>