avoid deprecation warnings for json_exception
This class is deprecated but we still have usages in routes.cc.

An attempt was made to silence them via #pragma GCC diagnostic ignored
"-Wdeprecated-declarations" around a using declaration, but that doesn't
work on clang at least (perhaps it does on gcc?).

Use a different approach by replacing it with a internal::to_json()
function inside the block, which also gives use a smaller target to
update in the future when we want to remove json_exception entirely.

Closes scylladb/seastar#3156