Add file::statat and a corresponding file_stat overload
Exposing the fstatat system call.
To be used for efficient directory listing
where, instead of calling file_stat for the full path,
the caller can use dir.statat(dirent.name) for each entry
which is more efficient as it doesn't require to construct
the full path and to look up the directory prefix on every call.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>