pollable_fd: Unfriend everything
There are four friend classes in it, but none really needs it. All of them
want to call get_fd() method, which can be safely made public (it's pollable
FD, why not export the fd value itself?). One exception is readable_eventfd
that tries to access pollable_fd_state pointer, it needs it to pass one into
reactor::readable() and pollable_fd itself has the readable() method that
does the same and is pubic.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>