seastar::test: Add exception interception to extract nested info
Fixes #3165

While not a seastar problem as such, but a weakness in the current boost::test,
any C++ exception failure in a test will only report the type and message of
the outer most exception, which in some cases, where nesting is used to
package the info, means in the end we have no idea why something failed.

This patch adds exception interception at the invocation of the test case,
normally done in execution_monitor.ipp, and similar to this, creates exception
info to be propagated to the test logging. However, unlike the boost impl,
we recurse the nested exceptions and add these types and messages as well.

Small unit test included.

Note: the actual (XML) logging of tests is limited to one type for an
exception causing failure, so we are limited to tracking additional info
in the "message" portion.

Note: similar to the original code in boost::test, this code avoid buffer
allocs to at least some degree (which can be argued if it really helps any).

v2:
* Renamed routine to repackage_exception_and_rethrow