dns: Use new datagram::get_buffers() API
The returned span of buffers is sequentially copied to the provided
buffer, similarly to how it was done with packet fragments. "Consumed"
buffers are trimmed to zero size, so in case the loop is called for
the 2nd time over the same datagram, it will no-op-loop over empty
buffers. Hopefully, DNS packets are not going to be heavily fragmented.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>