memcache: Use new datagram::get_buffers() API
When the datagram is received, the code grabs first 16 bytes from it
for the header, then converts the remainder into memory input_stream.

The new code converts buffers into input_stream instantly, and then
uses input_stream::read_exactly() to gram the necessary amount of bytes
into the header. This shouldn't be very expensive, when the necessary
size is available, it will just return the share()-d temporary_buffer.

This change, however, affects the way "connection" is initialized. Now
the request_id is not available construction-time, but since it's only
needed by the time response is sent, it's also fine.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>