io_queue: Track integrated-length for queue- and disk-queue- lengths
Each priority class currently calculates the number of queued requests
and the number of executing requests, and exports both as gauges. This
patch switches both counters to be integrated-length values and exports
both as counters.

Checkpoint happens -- for queued requests before dispatching, for
executing requests after dispatching. Thus the counter provides the info
about the lengths of queues scheduler and disk have to deal with,
respectively.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>