test: Improve resilience of netsed scheduling fairness test
The test runs a bunch of busyloops in different groups hierarchies and
checks that the consumed CPU is proporitonal to expectations, that are
based on shares values.

The test uses real-time clocks (because CPU scheduler can use no other
clocks currently) and sometimes fails on busy nodes that steal time from
test and don't let group accumulate counters in expected manner.

This patch shuffles the test so that bosyloops are run until they are
explicitly stopped, and the main fiber sleeps for some time, then checks
the accumulated counters. If the requirements are met the test passes,
otherwise loops are given more time (exponentially increasing) to run
and count. The test fails only if counters do not converge within several
minutes. Ideally it should happen much faster, within few hundreeds of
milliseconds.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>

Closes scylladb/seastar#3179