json: formatter: fix formatting of std:string_view
Currently, to_json of a std::string_view resolved to
the Range overload, and so it is printed as array of integers
rather than as a string.

For example: std::string_view("hello, world") is wrongly formatted
as [104,101,108,108,111,44,32,119,111,114,108,100]

The fix is to implement json formatting for std::string_view
rather than const sstring&, allowing the decay of seastar::string
or std::string to std::string_view via to implicit type conversion.

Fixes #3183

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>

sq